+++ NEWS.md
@@ -1,6 +1,8 @@
 # AzureDatalakeStore 0.0.0.9000
 
-* Added a `NEWS.md` file to track changes to the package.
+* Adds function `adls_url()` to compose the base URL.
+
+* Adds a `NEWS.md` file to track changes to the package.